Separate SpaceCardModalDemo into its functional components
