+++ pages/spaces/spaceCardDemo.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SpacesCard from "../../components/Spaces/Dashboard/SpacesCard";
+import SpaceCard from "../../components/Spaces/Dashboard/SpaceCard";
 import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
 
 const theme = createMuiTheme({
@@ -16,25 +16,25 @@ const theme = createMuiTheme({
   },
 });
 
-export default function SpacesCardDemo() {
+export default function SpaceCardDemo() {
   return (
     <ThemeProvider theme={theme}>
       <div className="flex flex-row space-x-5">
         <div className="cursor-pointer transform hover:scale-110 transition ease-out duration-200">
-          <SpacesCard
+          <SpaceCard
             spaceName="UW Math 2025"
             description="finals grind, upper years available in chat for help with past exams"
             headCount="17"
             music="lofi 2"
           />
         </div>
-        <SpacesCard
+        <SpaceCard
           spaceName="Capstone Grind '25"
           description="writing your report, making your presentation, setting up data"
           headCount="23"
           music="cafe beats eng edition F21"
         />
-        <SpacesCard
+        <SpaceCard
           spaceName="UW Math 2025"
           description="3rd and 4th years offering help in MSCI, GENE, MATH, and CS"
           headCount="8"
